+++ src/hiro_graph_client/__init__.py
@@ -9,8 +9,8 @@ from hiro_graph_client.authclient import HiroAuth
 from hiro_graph_client.batchclient import HiroGraphBatch, SessionData, AbstractIOCarrier, BasicFileIOCarrier, \
     SourceValueError, HiroResultCallback
 from hiro_graph_client.client import HiroGraph
-from hiro_graph_client.clientlib import HiroApiHandler, PasswordAuthTokenHandler, FixedTokenHandler, \
-    EnvironmentTokenHandler, AbstractTokenHandler, AuthenticationTokenError, FixedTokenError, TokenUnauthorizedError, \
+from hiro_graph_client.clientlib import AbstractTokenApiHandler, AuthenticationTokenError, FixedTokenError, \
+    TokenUnauthorizedError, PasswordAuthTokenApiHandler, FixedTokenApiHandler, EnvironmentTokenApiHandler, \
     accept_all_certs
 
 this_directory = path.abspath(path.dirname(__file__))
@@ -18,10 +18,10 @@ with open(path.join(this_directory, 'VERSION'), encoding='utf-8') as f:
     __version__ = f.read().strip()
 
 __all__ = [
-    'HiroApiHandler', 'HiroGraph', 'HiroAuth', 'HiroApp', 'HiroGraphBatch', 'SessionData', 'HiroResultCallback',
-    'AbstractTokenHandler', 'PasswordAuthTokenHandler', 'FixedTokenHandler', 'EnvironmentTokenHandler',
-    'AuthenticationTokenError', 'FixedTokenError', 'TokenUnauthorizedError', 'AbstractIOCarrier', 'BasicFileIOCarrier',
-    'SourceValueError', 'accept_all_certs', '__version__'
+    'HiroGraph', 'HiroAuth', 'HiroApp', 'HiroGraphBatch', 'SessionData',
+    'HiroResultCallback', 'AbstractTokenApiHandler', 'PasswordAuthTokenApiHandler', 'FixedTokenApiHandler',
+    'EnvironmentTokenApiHandler', 'AuthenticationTokenError', 'FixedTokenError', 'TokenUnauthorizedError',
+    'AbstractIOCarrier', 'BasicFileIOCarrier', 'SourceValueError', 'accept_all_certs', '__version__'
 ]
 
 site.addsitedir(this_directory)