+++ src/hiro_graph_client/authclient.py
@@ -2,36 +2,23 @@
 
 from typing import Any, Iterator
 
-from hiro_graph_client.clientlib import HiroApiHandler, AuthenticatedAPI, AbstractTokenHandler
+from hiro_graph_client.clientlib import AbstractHandledAPI, AbstractTokenApiHandler
 
 
-class HiroAuth(AuthenticatedAPI):
+class HiroAuth(AbstractHandledAPI):
     """
     Python implementation for accessing the HIRO Auth REST API.
     See https://core.arago.co/help/specs/?url=definitions/auth.yaml
     """
 
-    def __init__(self,
-                 api_handler: HiroApiHandler = None,
-                 endpoint: str = None,
-                 token_handler: AbstractTokenHandler = None,
-                 raise_exceptions: bool = False,
-                 proxies: dict = None):
+    def __init__(self, api_handler: AbstractTokenApiHandler):
         """
         Constructor
 
-        :param api_handler: Instance of a version handler that contains the current API endpoints.
-        :param endpoint: Full url for auth API. Overrides endpoints taken from *api_handler*.
-        :param token_handler: External token handler. An internal one is created when this is unset.
-        :param raise_exceptions: Raise exceptions on HTTP status codes that denote an error. Default is False
-        :param proxies: Proxy configuration for *requests*. Default is None.
+        :param api_handler: External API handler.
         """
         super().__init__(api_name='auth',
-                         api_handler=api_handler,
-                         endpoint=endpoint,
-                         token_handler=token_handler,
-                         raise_exceptions=raise_exceptions,
-                         proxies=proxies)
+                         api_handler=api_handler)
 
     ###############################################################################################################
     # REST API operations against the auth API
@@ -43,7 +30,7 @@ class HiroAuth(AuthenticatedAPI):
 
         :return: The result payload
         """
-        url = self._endpoint + '/me/account'
+        url = self.endpoint + '/me/account'
         return self.get(url)
 
     def get_avatar(self) -> Iterator[bytes]:
@@ -52,7 +39,7 @@ class HiroAuth(AuthenticatedAPI):
 
         :return: The result payload yields over binary data. Complete binary payload is an image/png.
         """
-        url = self._endpoint + '/me/avatar'
+        url = self.endpoint + '/me/avatar'
         yield self.get_binary(url, accept='image/png')
 
     def put_avatar(self, data: Any) -> dict:
@@ -62,7 +49,7 @@ class HiroAuth(AuthenticatedAPI):
         :param data: Binary data for image/png of avatar.
         :return: The result payload
         """
-        url = self._endpoint + '/me/avatar'
+        url = self.endpoint + '/me/avatar'
         return self.put_binary(url, data, content_type='image/png')
 
     def change_password(self, old_password: str, new_password: str) -> dict:
@@ -73,7 +60,7 @@ class HiroAuth(AuthenticatedAPI):
         :param new_password: The new password.
         :return: The result payload
         """
-        url = self._endpoint + '/me/password'
+        url = self.endpoint + '/me/password'
 
         data = {
             "oldPassword": old_password,
@@ -88,7 +75,7 @@ class HiroAuth(AuthenticatedAPI):
 
         :return: The result payload
         """
-        url = self._endpoint + '/me/profile'
+        url = self.endpoint + '/me/profile'
         return self.get(url)
 
     def post_profile(self, data: dict) -> dict:
@@ -99,7 +86,7 @@ class HiroAuth(AuthenticatedAPI):
                See https://core.arago.co/help/specs/?url=definitions/auth.yaml#/[Me]_Identity/post_me_profile
         :return: The result payload
         """
-        url = self._endpoint + '/me/profile'
+        url = self.endpoint + '/me/profile'
         return self.post(url, data)
 
     def get_roles(self) -> dict:
@@ -108,7 +95,7 @@ class HiroAuth(AuthenticatedAPI):
 
         :return: The result payload
         """
-        url = self._endpoint + '/me/roles'
+        url = self.endpoint + '/me/roles'
         return self.get(url)
 
     def get_teams(self) -> dict:
@@ -117,5 +104,5 @@ class HiroAuth(AuthenticatedAPI):
 
         :return: The result payload
         """
-        url = self._endpoint + '/me/teams'
+        url = self.endpoint + '/me/teams'
         return self.get(url)